+++ lib/src/database/weapons/overpower.dart
@@ -1,0 +1,8 @@
+import 'package:swlegion/swlegion.dart';
+
+final overpower = Weapon.melee(
+  name: 'Overpower',
+  dice: {
+    AttackDice.red: 4,
+  },
+);